FrontUI: Double to BigDecimal refactoring for amount

+++ frontui/src/main/java/ru/yandex/practicum/frontui/dto/AccountDto.java
@@ -1,4 +1,6 @@
 package ru.yandex.practicum.frontui.dto;
 
-public record AccountDto(Long id, String currency, Double balance) {
+import java.math.BigDecimal;
+
+public record AccountDto(Long id, String currency, BigDecimal balance) {
 }
